fix: fallback to the user directory if no a valid global config directory
In the not supported for DSG_APP_DATA environment, we can't get a valid
directory for the global config, then falied when saving data. In order
not to lose data, we can only fallback to storing it to the user
directory.

Also, allows to specify a fallback directory for DSG_APP_DATA on
compile, in order to never fallback to the user directory.

Log: