fix: calling delay in DDBusSender
When creating QAbstractInterface, Qt will try to invoke instrospection
into the DBus path, this is sometimes costing when instrospection
is not ready for that path. Change the implementation of DDBusSender.
Don't use QDBusInterface. Instead, create a QDBusMessage and invoke it
directly. This is sensible because DDBusSender is just a wrapper like
QDBusInterface, it shouldn't rely on QDBusInterface.

Issue: https://github.com/linuxdeepin/developer-center/issues/4415
Log: fix calling delay in DDBusSender